angular fix protractor tests
